Do not list self as seed when joining a non-empty cluster.
When a node joins a non-empty cluster, it should not list itself as
seed.
While by default the node will still try to contact other seeds and
stream data if it discovers them (controlled by auto_bootstrap option,
which defaults to true), if other seeds are not reachable
within ring_timeout the node listed as seed will go on serving
reads and writes without prior streaming.

This violates LWT linearisability.